move distillation runner config to correct file
